# Optimize Hpack and AsciiString hashcode and equals